+++ app/lib/shared/versions.dart
@@ -22,7 +22,7 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// when the version switch happens.
 const acceptedRuntimeVersions = <String>[
   // The current [runtimeVersion].
-  '2022.09.01',
+  '2022.09.08',
   // Fallback runtime versions.
   '2022.08.31',
   '2022.08.26',
@@ -55,9 +55,9 @@ bool shouldGCVersion(String version) =>
 // keep in-sync with SDK version in .mono_repo.yml and Dockerfile
 final String runtimeSdkVersion = '2.18.0';
 final String toolStableDartSdkVersion = '2.18.0';
-final String toolStableFlutterSdkVersion = '3.3.0';
+final String toolStableFlutterSdkVersion = '3.3.1';
 final String toolPreviewDartSdkVersion = '2.18.0';
-final String toolPreviewFlutterSdkVersion = '3.3.0';
+final String toolPreviewFlutterSdkVersion = '3.3.1';
 
 final semanticToolStableDartSdkVersion =
     Version.parse(toolStableDartSdkVersion);